test: split oversized test files by responsibility (#61)
Closes #61.

Split tests/test_refactored_modules.py (492 lines, 52 tests) into two
focused files:
- tests/test_refactored_core.py (370 lines): validator, dxf_builder,
  stream_router, control_loops, notes
- tests/test_refactored_generator.py (158 lines): generator orchestration
  helpers and sheet_layout

Both files are now well under the 500-line budget. Test count preserved
(52 tests, all passing). No production code changes.

Fleet old-issue triage — wave 27.
